requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pymongo[srv]
certifi
pandas
openpyxl
xlsxwriter
python-dotenv
passlib[bcrypt]
bcrypt==4.0.1
starlette
itsdangerous


ok can you give a code inject.py to add some values to the database to check the working because we dont have the app yet so ok just some values to test the things i wil give you some creatntioals wait 

_id

696613a87cf8f8cdf4e9b1d5

username

"Prajwal Ganiga"

email

"prajwalganiga06@gmail.com"

password_hash

"$2b$12$1ZLjlFgrewgImPE1dB9rCuskoZA2Cro.kyb6wfLiDXtmabOPYZrMm" [prajwal]

role

"super_admin"

reset_otp

"990676" 

create 4 customers and a driver and make the things work for this admin use the location of Mangalore,Padil in Karnataka dakshina kannada ok use some loctions of there 